feat: added some tests
