+++ info.plist
@@ -1,89 +1,135 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-  <dict>
-    <key>bundleid</key>
-    <string>com.alfredapp.notion</string>
-    <key>category</key>
-    <string>Productivity</string>
-    <key>connections</key>
-    <dict/>
-    <key>createdby</key>
-    <string>Anand Rajaram</string>
-    <key>description</key>
-    <string>An Alfred workflow to quickly add tasks to Notion</string>
-    <key>disabled</key>
-    <false/>
-    <key>name</key>
-    <string>Notion Alfred</string>
-    <key>objects</key>
-    <array>
-      <dict>
-        <key>config</key>
-        <dict>
-          <key>alfredfiltersresults</key>
-          <false/>
-          <key>alfredfiltersresultsmatchmode</key>
-          <integer>0</integer>
-          <key>argumenttreatemptyqueryasnil</key>
-          <true/>
-          <key>argumenttrimmode</key>
-          <integer>0</integer>
-          <key>argumenttype</key>
-          <integer>0</integer>
-          <key>escaping</key>
-          <integer>102</integer>
-          <key>keyword</key>
-          <string>nta</string>
-          <key>queuedelaycustom</key>
-          <integer>3</integer>
-          <key>queuedelayimmediatelyinitially</key>
-          <true/>
-          <key>queuedelaymode</key>
-          <integer>0</integer>
-          <key>queuemode</key>
-          <integer>1</integer>
-          <key>runningsubtext</key>
-          <string>Please Wait</string>
-          <key>script</key>
-          <string>./node_modules/.bin/run-node index.js "$1"
+<dict>
+	<key>bundleid</key>
+	<string>com.alfredapp.notion</string>
+	<key>category</key>
+	<string>Productivity</string>
+	<key>connections</key>
+	<dict>
+		<key>05CCF908-66B0-406D-8A1C-7C71603CA4F8</key>
+		<array/>
+		<key>AD7A029E-AF29-4609-AED5-CEFBE2215643</key>
+		<array>
+			<dict>
+				<key>destinationuid</key>
+				<string>05CCF908-66B0-406D-8A1C-7C71603CA4F8</string>
+				<key>modifiers</key>
+				<integer>0</integer>
+				<key>modifiersubtext</key>
+				<string></string>
+				<key>vitoclose</key>
+				<false/>
+			</dict>
+		</array>
+	</dict>
+	<key>createdby</key>
+	<string>Anand Rajaram</string>
+	<key>description</key>
+	<string>An Alfred workflow to quickly add tasks to Notion</string>
+	<key>disabled</key>
+	<false/>
+	<key>name</key>
+	<string>Notion Alfred</string>
+	<key>objects</key>
+	<array>
+		<dict>
+			<key>config</key>
+			<dict>
+				<key>alfredfiltersresults</key>
+				<false/>
+				<key>alfredfiltersresultsmatchmode</key>
+				<integer>0</integer>
+				<key>argumenttreatemptyqueryasnil</key>
+				<true/>
+				<key>argumenttrimmode</key>
+				<integer>0</integer>
+				<key>argumenttype</key>
+				<integer>0</integer>
+				<key>escaping</key>
+				<integer>102</integer>
+				<key>keyword</key>
+				<string>nta</string>
+				<key>queuedelaycustom</key>
+				<integer>3</integer>
+				<key>queuedelayimmediatelyinitially</key>
+				<true/>
+				<key>queuedelaymode</key>
+				<integer>1</integer>
+				<key>queuemode</key>
+				<integer>1</integer>
+				<key>runningsubtext</key>
+				<string>Please Wait</string>
+				<key>script</key>
+				<string>./node_modules/.bin/run-node index.js "$1"
 </string>
-          <key>scriptargtype</key>
-          <integer>1</integer>
-          <key>scriptfile</key>
-          <string/>
-          <key>subtext</key>
-          <string>Update available: 1.0.0 → 1.1.0. Run `npm install -g alfred-notion`</string>
-          <key>title</key>
-          <string>Enter your task</string>
-          <key>type</key>
-          <integer>0</integer>
-          <key>withspace</key>
-          <true/>
-        </dict>
-        <key>type</key>
-        <string>alfred.workflow.input.scriptfilter</string>
-        <key>uid</key>
-        <string>AD7A029E-AF29-4609-AED5-CEFBE2215643</string>
-        <key>version</key>
-        <integer>3</integer>
-      </dict>
-    </array>
-    <key>readme</key>
-    <string/>
-    <key>uidata</key>
-    <dict>
-      <key>AD7A029E-AF29-4609-AED5-CEFBE2215643</key>
-      <dict>
-        <key>xpos</key>
-        <integer>150</integer>
-        <key>ypos</key>
-        <integer>180</integer>
-      </dict>
-    </dict>
-    <key>version</key>
-    <string/>
-    <key>webaddress</key>
-    <string>https://anandrajaram.dev</string>
-  </dict>
+				<key>scriptargtype</key>
+				<integer>1</integer>
+				<key>scriptfile</key>
+				<string></string>
+				<key>subtext</key>
+				<string></string>
+				<key>title</key>
+				<string>Enter your task</string>
+				<key>type</key>
+				<integer>0</integer>
+				<key>withspace</key>
+				<true/>
+			</dict>
+			<key>type</key>
+			<string>alfred.workflow.input.scriptfilter</string>
+			<key>uid</key>
+			<string>AD7A029E-AF29-4609-AED5-CEFBE2215643</string>
+			<key>version</key>
+			<integer>3</integer>
+		</dict>
+		<dict>
+			<key>config</key>
+			<dict>
+				<key>concurrently</key>
+				<false/>
+				<key>escaping</key>
+				<integer>102</integer>
+				<key>script</key>
+				<string>./node_modules/.bin/run-node notion.js "$1"</string>
+				<key>scriptargtype</key>
+				<integer>1</integer>
+				<key>scriptfile</key>
+				<string></string>
+				<key>type</key>
+				<integer>0</integer>
+			</dict>
+			<key>type</key>
+			<string>alfred.workflow.action.script</string>
+			<key>uid</key>
+			<string>05CCF908-66B0-406D-8A1C-7C71603CA4F8</string>
+			<key>version</key>
+			<integer>2</integer>
+		</dict>
+	</array>
+	<key>readme</key>
+	<string></string>
+	<key>uidata</key>
+	<dict>
+		<key>05CCF908-66B0-406D-8A1C-7C71603CA4F8</key>
+		<dict>
+			<key>xpos</key>
+			<integer>320</integer>
+			<key>ypos</key>
+			<integer>180</integer>
+		</dict>
+		<key>AD7A029E-AF29-4609-AED5-CEFBE2215643</key>
+		<dict>
+			<key>xpos</key>
+			<integer>150</integer>
+			<key>ypos</key>
+			<integer>180</integer>
+		</dict>
+	</dict>
+	<key>version</key>
+	<string></string>
+	<key>webaddress</key>
+	<string>https://anandrajaram.dev</string>
+</dict>
 </plist>
