+++ package.json
@@ -12,7 +12,11 @@
     "type": "git",
     "url": "git+ssh://git@github.com/anandrajaram21/alfred-notion.git"
   },
-  "author": "Anand Rajaram",
+  "author": {
+    "name": "Anand Rajaram",
+    "email": "anand@anandrajaram.dev",
+    "url": "anandrajaram.dev"
+  },
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/anandrajaram21/alfred-notion/issues"
@@ -20,6 +24,7 @@
   "homepage": "https://github.com/anandrajaram21/alfred-notion#readme",
   "dependencies": {
     "@notionhq/client": "^0.3.2",
+    "alfred-notifier": "^0.2.3",
     "alfy": "^0.12.1",
     "chrono-node": "^2.3.1",
     "dotenv": "^10.0.0"
